+++ A
@@ -1,0 +1,2 @@
+Nothing specified, nothing added.
+Maybe you wanted to say 'git add .'?
