+++ src/components/Music.vue
@@ -1,0 +1,15 @@
+<template>
+  <el-card>
+    <img
+      ="https://pbs.twimg.com/profile_images/1404769598562455554/BOFS5GUG_400x400.png"
+    />
+  </el-card>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
+  setup() {},
+})
+</script>
